update : NetworkSettings.kt 폴더 이동, NoConnectedDialog.kt activity -> context로 변경에 따른 코드 refactoring
